software/requirements_all.txt
pyrtlsdr
hashlib